api/requirements.txt
flask
numpy
scikit-learn
pandas
numpy
requests
mysql-connector-python